+++ Configuration/Backend/Modules.php
@@ -2,34 +2,10 @@
 
 declare(strict_types=1);
 
+use Ppl\PplDeeplV3Translate\Controller\BackendFrontendAccessController;
 use Ppl\PplDeeplV3Translate\Controller\BackendTranslationController;
-use Ppl\PplDeeplV3Translate\Controller\BackendConfigurationController;
 
 return [
-    'ppl_deepl_v3' => [
-        'position' => ['after' => 'system'],
-        'iconIdentifier' => 'module-ppl-deepl-v3',
-        'labels' => [
-            'title' => 'LLL:EXT:ppl_deepl_v3_translate/Resources/Private/Language/locallang.xlf:module.root.title',
-            'shortDescription' => 'LLL:EXT:ppl_deepl_v3_translate/Resources/Private/Language/locallang.xlf:module.root.description',
-        ],
-    ],
-    'ppl_deepl_v3_configuration' => [
-        'parent' => 'ppl_deepl_v3',
-        'position' => ['before' => '*'],
-        'access' => 'user',
-        'path' => '/module/ppl-deepl-v3/configuration',
-        'iconIdentifier' => 'module-ppl-deepl-v3-configuration',
-        'labels' => [
-            'title' => 'LLL:EXT:ppl_deepl_v3_translate/Resources/Private/Language/locallang.xlf:module.configuration.title',
-            'shortDescription' => 'LLL:EXT:ppl_deepl_v3_translate/Resources/Private/Language/locallang.xlf:module.configuration.description.v3',
-        ],
-        'routes' => [
-            '_default' => [
-                'target' => BackendConfigurationController::class . '::handleRequest',
-            ],
-        ],
-    ],
     'ppl_deepl_v3_translation' => [
         'parent' => 'ppl_deepl_v3',
         'position' => ['after' => 'ppl_deepl_v3_configuration'],
@@ -59,6 +35,27 @@ return [
         'routes' => [
             '_default' => [
                 'target' => BackendTranslationController::class . '::handleRequest',
+            ],
+            'download' => [
+                'path' => '/download',
+                'methods' => ['GET'],
+                'target' => BackendTranslationController::class . '::downloadAction',
+            ],
+        ],
+    ],
+    'ppl_deepl_v3_frontend_access' => [
+        'parent' => 'ppl_deepl_v3',
+        'position' => ['after' => 'ppl_deepl_v3_file_translation'],
+        'access' => 'user',
+        'path' => '/module/ppl-deepl-v3/frontend-access',
+        'iconIdentifier' => 'module-ppl-deepl-v3-configuration',
+        'labels' => [
+            'title' => 'LLL:EXT:ppl_deepl_v3_translate/Resources/Private/Language/locallang.xlf:config.frontendAccess',
+            'shortDescription' => 'LLL:EXT:ppl_deepl_v3_translate/Resources/Private/Language/locallang.xlf:backend.frontendLoginApplies',
+        ],
+        'routes' => [
+            '_default' => [
+                'target' => BackendFrontendAccessController::class . '::handleRequest',
             ],
         ],
     ],